chore(blockchain-link): move solana types
