+++ packages/blockchain-link-utils/src/solana.ts
@@ -2,6 +2,14 @@ import { A, D, F, pipe } from '@mobily/ts-belt';
 
 import { Target, TokenTransfer, Transaction } from '@trezor/blockchain-link-types/src';
 import type { StakeType, TokenInfo, TokenStandard } from '@trezor/blockchain-link-types/src';
+import {
+    SolanaTokenAccountInfo,
+    TokenDetailByMint,
+} from '@trezor/blockchain-link-types/src/solana';
+import { isCodesignBuild } from '@trezor/env-utils';
+import { arrayPartition } from '@trezor/utils';
+import { BigNumber } from '@trezor/utils/src/bigNumber';
+
 import type {
     AccountInfo,
     Address,
@@ -9,14 +17,8 @@ import type {
     ParsedInstruction,
     ParsedTransactionWithMeta,
     PartiallyDecodedInstruction,
-    SolanaTokenAccountInfo,
     SolanaValidParsedTxWithMeta,
-    TokenDetailByMint,
-} from '@trezor/blockchain-link-types/src/solana';
-import { isCodesignBuild } from '@trezor/env-utils';
-import { arrayPartition } from '@trezor/utils';
-import { BigNumber } from '@trezor/utils/src/bigNumber';
-
+} from './solana-types';
 import { formatTokenSymbol } from './utils';
 
 export type ApiTokenAccount = {
